Update 03-conda-basics.md
Added a line about requiring "bash". I stumbled over this since I (and quite a few others) are using "tcsh" on Linux and Conda doesn't play well. I tried a few hacks to make this work with tcsh but gave up on it. Don't know how this works on Windows/Mac

I was trying to figure out how to execute a python script from the shell and use a specific command. After looking for quite a while and finding only convoluted answers, I finally found this. Just replace the first line in the python script with the path to the python instance of the environment you use. Maybe useful for others (unless trivial)
#!/home/axel/anaconda3/envs/my_env37/bin/python
Found the solution here: 
From: https://stackoverflow.com/questions/51469569/conda-and-python-shell-scripts